+++ tutorial/create_images.sh
@@ -1,59 +1,50 @@
 #!/bin/sh
 
-# simple script to run tutorials, create images, and to 
-# install them in the libfaudes doxygen tree
+# run relevant turtorials and process images for reference and doxugen
 
 # configure
-LIBFAUDES=../
+LIBFAUDES=..
+LUAFAUDES=$LIBFAUDES/bin/luafaudes
+GEN2DOT=$LIBFAUDES/bin/gen2dot
+DSTDIR=../src/doxygen/faudes_images
 
-DOTWRITE=$LIBFAUDES/bin/gen2dot
-DOTEXEC=dot
-CONVERT=convert
-DSTDIR=$LIBFAUDES/src/doxygen/faudes_images
-GEN2REF=$LIBFAUDES/tools/gen2ref/gen2ref.pl
+# sanity check
+if [ ! -d $DSTDIR ]; then
+    echo "error: the current path appears not to be a libFAUDES tutorial"
+    return
+fi
+if [ ! -f $LIBFAUDES/src/registry/cfl_definitions.rti ]; then
+    echo "error: the current path appears not to be a libFAUDES tutorial"
+    return
+fi
+
 
 # advertise
 echo ======================================================
-echo ===  running tutorials ===============================
+echo ===  running tutorials to generate .gen files ========
 echo ======================================================
 
-rm tmp_*
+rm -f tmp_*
 
 ./3_functions
 ./1_generator
 
+rm tmp_languagesEmpty_result*
 cp data/simplemachine.gen tmp_simplemachine.gen
 
-# advertise
-echo ======================================================
-echo ===  converting gen to png/svg/fref ==================
-echo ======================================================
-
-
-# loop all .gen files for graphics
-for FILE in tmp_*.gen ; do
-  BASE=$(basename $FILE .gen)
-  echo ============= processing $BASE
-  $DOTWRITE $FILE
-  $DOTEXEC -Tsvg -Gbgcolor=transparent -Gsize=10,10 $BASE.dot -o $BASE.svg
-  $CONVERT -background none $BASE.svg $BASE.png
-done;
-
-# loop all .gen files for fref page
-for FILE in tmp_*.gen ; do
-  BASE=$(basename $FILE .gen)
-  echo ============= processing $BASE
-  $GEN2REF $BASE.gen > $BASE.fref
-done;
+# pass on to common image generation
+. ../tools/imgproc/imgproc.sh
 
 # advertise
 echo ======================================================
-echo ===  copy do doc =====================================
+echo ===  copy to doc =====================================
 echo ======================================================
 
-rm $DSTDIR/tmp_*
-cp -v tmp_*.png $DSTDIR
-cp -v tmp_*.svg $DSTDIR
-cp -v tmp_*.fref $DSTDIR
+#rm $DSTDIR/tmp_*
+#cp -v tmp_*.png $DSTDIR
+#cp -v tmp_*.svg $DSTDIR
+#cp -v tmp_*.fref $DSTDIR
+
+
